requirements.txt
requests
pandas
plotly
jinja2
python-dateutil
python-dotenv